Move function class to interpreter_runtime
